context_eval/requirements.txt
pandas
numpy
scipy
openai
python-dotenv
praw
openpyxl
tqdm
